+++ test/integration/shpec/run_exit_status_preserved.sh
@@ -5,7 +5,7 @@ describe "ide command: run preserves exit status"
   describe 'when IDE_DRIVER="docker"'
     describe 'when exit 164 in docker container'
       it "returns 164"
-        message=$(cd test/docker/dummyide-usage && IDE_LOG_LEVEL=debug ${IDE_PATH} 'echo abc && exit 164')
+        message=$(cd test/docker/example-ide-usage && IDE_LOG_LEVEL=debug ${IDE_PATH} 'echo abc && exit 164')
         exit_status="$?"
         it "exits with status 164"
           assert equal "$exit_status" "164"